requirements.txt
flask
azure-storage-blob
gunicorn
python-dotenv
